:sparkles: Add GameTime class to keep track of time in the game
The GameTime is used to make sure that all processes on the game run at the exact same time regardless of what device you are on.

The class has a tick method which returns a tuple of the total time and the elapsed time. There is also a getter method to get the elapsed_game_time