:sparkles: Add ThirdPersonController class to update the movement of the camera to follow the player
This class is attached to a camera object.The class updates the camera position to follow the target.The class makes sure that the target is always in the center of the screen.

The RendererManager had to be updated to draw the position of the objects relative to where the camera was now positioned.